Edison mods for libmraa
Check if SPI io pointers holding are valid data before using them.

libmraa needs to set pin modes via nodes in debugfs. This patch
adds a feature rc file to mount debugfs. This rc file is on
system partition so we need to put it to a trigger later than
mounting system.

pull in glob() from the avahi BSP project.
  bsd licensed code

Change-Id: I7afd8efb4335ef703ab447ed20d78796ea57c9b4
Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>