edison: add gpt_ini2bin tool
The gpt_ini2bin.py script creates a serialized format of part of
the gpt entries table. When the generated binary is flashed to the
device under the 'gpt' target, the device will fully create the
whole gpt and flash it to storage.

The partial gpt representation created by this tool is agnostic to
some device characteristics like block size (sizes are specified
in units of MiB) and total device size. A single partition can be
given size -1 and will expand to fit the remaining storage after
taking space for the concretely sized partitions.

Change-Id: I80f5796dad008912eae53a4dcfdaccfaac125e01
Signed-off-by: Scott D Phillips <scott.d.phillips@intel.com>