[libupm] Edison mods for libupm
The existing hardcoded logic enables i2c bypass mode for AK8975.
But this will makes accelerometer disappear on I2C bus. We add
a new member as a switch and disable bypass by default.

Most of source files are with extension cxx but a few are named
with cpp. Build system cannot handle both of them in a single
project:
  This patch adds all modules having cxx files into build.

  Several modules with cpp files are not added this time, for
  the sake of later update from upm git hub until they are really
  needed.

  Module nrf8001 have both cpp and cxx, we have to rename them
  this time.

Change-Id: Icdf40f0784f8a3485225a6770d3966867613a2e2
Author: Jianxun Zhang <jianxun.zhang@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>