light: add support for Minnowboard Turbot
Minnowboard Turbot has the D2 LED under GPIO control.
Associate the Notification light with D2.
Refactor the code to identify board versions using mraa_get_platform_version.
The D2 LED present on Minnowboard Turbot turns on if
the GPIO is set to 0 and off otherwise.
Apply value inversion so that 1 means that the light
is turned on.

BUG=none
Change-Id: I5c097da6969de8780e26266b5f74d2dd6bb90ae2
Tracked-On: https://jira01.devtools.intel.com/browse/BP-243
Signed-off-by: Constantin Musca <constantin.musca@intel.com>