libmraa: minnow: include Turbot additions
The pinout for the Low Speed Expansion on the MinnowBoard
Turbot is the same as the MinnowBoard MAX A2 design, with
the exception of pin 26. Pin 26 was changed to provide an MCLK
reference clock for I2S.
In addition, the D2 LED is now under GPIO control.
Add support for these changes.

BUG=none
Change-Id: Id3153af731a32afbc7b1b69c8ccab909f48ef4b0
Tracked-On: https://jira01.devtools.intel.com/browse/BP-244
Signed-off-by: Constantin Musca <constantin.musca@intel.com>