sensors: make the HAL thread safe
According to the Sensors HAL interface documentation, the implementation
must be thread safe and allow functions defined in sensors.h to be called
from different threads at the same time.
Use a mutex to synchronize the HAL. In addition, initialize the acquisition
pipes with the O_NONBLOCK flag to return immediately in case another tread
which calls poll() reads the data.
Put the read pipe call in a loop as the read system call might return a number
less than the number of requested bytes.

BUG=none
Change-Id: I6ee3f8e0c2ba19942da8b82da1660c9163c5d237
Tracked-On: https://jira01.devtools.intel.com/browse/BP-247
Signed-off-by: Constantin Musca <constantin.musca@intel.com>